move ports, stop hardcoding jinja dir
